% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectorize_comtrade_64.R
\name{sectorize_comtrade_64}
\alias{sectorize_comtrade_64}
\title{sectorize comtrade to 64 sectors}
\usage{
sectorize_comtrade_64(path)
}
\arguments{
\item{path, }{replace with directory where Comtrade bulk downloads are located}
}
\description{
maps and aggregates product level trade data to mrio sectors
}
